Squashed commits

+++ tests/language/vm/licm_constant_redefinition_vm_test.dart
@@ -14,7 +14,7 @@ class X {
 loop(val) {
   var sum = 0;
   for (var i = 0; i < 10; i++) {
-    sum += val.length;
+    sum += val.length as int;
   }
   return sum;
 }
